// src/app/model/user-model.ts
export class UserModel {
  userId: string="";
  password: string="";
  username: string="";
  firstname: string="";
  lastname: string="";
  email: string="";
  secondemail: string="";
  phone: string="";
  secondphone: string="";
  status: string="";
  macaddress: string="";
  ipaddress: string="";
  accounttype: string="";


  dob: string="";
  address: string="";
  district: string="";

  zone: string="";

  state: string="";
  city: string="";
  zip: string="";
  companyname: string="";
  companyaddress: string="";
  monthlysalary: string="";
  monthlyexpense: string="";
  customerimage: any="";
}